FEAT: 값 타입 실전 예제
1. 값 타입 실전 예제
